dom/src/threads/test/test_scopeOnerror.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  const Cc = Components.classes;
  const Ci = Components.interfaces;
  const Cr = Components.results;

  var errors = [
    "This exception should make it to onerror",
    "This error should not make it back out"
  ];

  var worker = new Worker("scopeOnError_worker.js");

  worker.onmessage = function(event) {
    ok(event.data == "Done");

    if (consoleListener) {
      netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
      consoleService.unregisterListener(consoleListener);
      consoleService.reset();
      consoleListener = null;
    }

    SimpleTest.finish();
  }

  worker.onerror = function(event) {
    is(event.target, worker);
    is(event.message,
       "uncaught exception: This exception should make it to onerror");
    event.preventDefault();
  };

  var consoleListener = {
    observe: function(message) {
      ok(false, "An error made it to the console: " + message);

      if (consoleListener) {
        netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
        consoleService.unregisterListener(consoleListener);
        consoleService.reset();
        consoleListener = null;
      }

      SimpleTest.finish();
    },

    QueryInterface: function(iid) {
      netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
      if (iid.equals(Ci.nsIConsoleListener) ||
          iid.equals(Ci.nsISupports)) {
        return this;
      }
      throw Cr.NS_NOINTERFACE;
    }
  };

  var consoleService =
    Cc["@mozilla.org/consoleservice;1"].getService(Ci.nsIConsoleService);

  consoleService.reset();
  consoleService.registerListener(consoleListener);

  for (var index in errors) {
    worker.postMessage(errors[index]);
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
